Solved a problem with multiple calls to a coroutine
Well, thi is little bit long. But, here I go.

This repo solve 2 main problems: Multiple calls to API when should be only once and wait to call the rest shit and multiple instances of caching, when should be only one.

Solve multiple instances of caching was really easy, just need to create a volatine instance of the repo and just re-call the instance after the first creation.

Solve the multiple API calls instead just wait for the first fucking call, was little bit tricky, 'coz I had to read/remember a tons of stuff about parallelism and related-with.

I think, was worth it do this mini-problem recreation.